Auth/backend/requirements.txt
fastapi[standard]
uvicorn
pydantic
sqlalchemy
databases
python-jose[cryptography]
passlib[bcrypt]
PyJWT
python-decouple
python-multipart
